refactor(suite-native): move backup and passphrase cards to components directory
