+++ suite-native/module-device-settings/src/components/PassphraseCard.tsx
